requirements.txt
pyjwt>=1.7.1
ujson>=1.35